Social Media Analyser/backend/requirements.txt
fastapi
uvicorn
requests
pandas
psycopg2-binary
python-dotenv
sqlalchemy
